pedidos/requirements.txt
Flask==2.3.3
redis==4.5.1
requests==2.28.1
mysql-connector-python==8.0.29